Removing Object checking from requests
Since #all_associations_and_attributes already return the reference instead of
objects, it is not necessary to check if the param is an object or a reference.
There are some cases (like the ones removed here) that the mentioned method
doesn't apply, but in this cases is more a responsability of the developer to
check then of the lib itself.

In both cases it make sense to not check it anymore and thats why this commit
removes them.