Moving #create_server logic from request to model
The attribute #template_name doesn't exist anymore since it value was
making the save method to have 2 different behaviors.

Now, to create a vm from scratch, you can simple call save on a new vm record.
And to create from a template, it is necessary to call all the methods (clone, provision, start)
on the template object. This change make it a bit more verbose but gives more control
over the flow, like use a cdrom as install-repository without the need to have another template.